+++ mullvad-cli/src/cmds/mod.rs
@@ -22,11 +22,11 @@ pub use self::connect::Connect;
 mod disconnect;
 pub use self::disconnect::Disconnect;
 
+mod dns;
+pub use self::dns::Dns;
+
 mod lan;
 pub use self::lan::Lan;
-
-mod custom_dns;
-pub use self::custom_dns::CustomDns;
 
 mod reconnect;
 pub use self::reconnect::Reconnect;
@@ -61,10 +61,9 @@ pub fn get_commands() -> HashMap<&'static str, Box<dyn Command>> {
         Box::new(Bridge),
         Box::new(Connect),
         Box::new(Disconnect),
+        Box::new(Dns),
         Box::new(Reconnect),
         Box::new(Lan),
-        #[cfg(not(target_os = "android"))]
-        Box::new(CustomDns),
         Box::new(Relay),
         Box::new(Reset),
         #[cfg(target_os = "linux")]